feat: move utils and exceptions to modules dir to allow for import in API/main.py
