+++ commands/config.py
@@ -3,8 +3,8 @@
 
 import sys
 
-import exceptions
-import utils
+import modules.exceptions as exceptions
+import modules.utils as utils
 
 
 def get_entered_config_key() -> str | None:
