+++ commands/diff.py
@@ -6,8 +6,8 @@ import difflib
 import sys
 from pathlib import Path
 
-import exceptions
-import utils
+import modules.exceptions as exceptions
+import modules.utils as utils
 from bs4 import BeautifulSoup
