+++ commands/switch.py
@@ -6,8 +6,8 @@ import shutil
 import sys
 from pathlib import Path
 
-import exceptions
-import utils
+import modules.exceptions as exceptions
+import modules.utils as utils
 
 
 def get_branch_to_switch() -> str | None:
